requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
discord.py==2.2.3
exceptiongroup==1.1.2
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1; python_version >= '3.7'
pathspec==0.11.1
platformdirs==3.8.1
pluggy==1.2.0
pyotp==2.8.0
pytest-docker==1.0.1
pytest-mock==3.11.1
pytest==7.4.0
pyyaml==6.0
requests==2.31.0
tomli==2.0.1
urllib3==2.0.3
yarl==1.9.2
